+++ src/app/app.module.ts
@@ -113,6 +113,7 @@ import { CodingDisplayComponent } from './lib/widgets/coding-display/coding-disp
 import { EnableWhenMethodComponent } from './lib/widgets/enable-when-method/enable-when-method.component';
 import { MetaProfileComponent } from './lib/widgets/meta-profile/meta-profile.component';
 import { UnsignedIntegerComponent } from './lib/widgets/unsigned-integer/unsigned-integer.component';
+import { ExtensionPrimitiveComponent } from './lib/widgets/extension-primitive/extension-primitive.component';
 
 import { environment } from '../environments/environment';
 import {IsDisabledPipe} from "./lib/pipes/is-disabled.pipe";
@@ -224,6 +225,7 @@ import {IsDisabledPipe} from "./lib/pipes/is-disabled.pipe";
     EnablewhenAnswerCodingComponent,
     EnableWhenSourceComponent,
     EntryFormatComponent,
+    ExtensionPrimitiveComponent,
     HelpTextComponent,
     ItemControlComponent,
     LabelRadioComponent,